requirements.txt
psycopg2-binary
Django>=2.1.3
drf-nested-routers==0.93.4
black
djangorestframework>=3.9.0
django-environ
django-leaflet
django-filter
django-cors-headers
python-dotenv
Pillow
requests
Twisted[tls,http2]
whitenoise
drf-yasg
djangorestframework_simplejwt
django-allauth==0.58.2
django-rest-auth==0.9.5
py_eureka_client
aio-pika